Fix project's structure
