--- helpers/constants.js
@@ -1,7 +1,0 @@
-const DB_PATH = "../db/contacts.json";
-const PORT = 3000;
-
-module.exports = {
-  DB_PATH,
-  PORT,
-};
