+++ helpers/index.js
@@ -1,9 +1,5 @@
-const constants = require('./constants');
-const contacts = require('./fetch-contacts');
-const HttpError = require('./http-error');
+const HttpError = require("./http-error");
 
 module.exports = {
-  constants,
-  contacts,
   HttpError,
-}
+};
