+++ server.js
@@ -1,6 +1,5 @@
 const app = require("./app");
-const { PORT } = require("./helpers/constants");
 
-app.listen(PORT, () => {
+app.listen(3000, () => {
   console.log("Server running. Use our API on port: 3000");
 });
